tele-repairs-bot/requirements.txt
fastapi==0.115.2
uvicorn==0.30.6
python-telegram-bot==21.6
gspread==6.1.2
google-auth==2.35.0
google-api-python-client==2.147.0
pydantic==2.9.2
